Search radius required if 'offline' is checked.
